requirements.txt
apache-airflow-providers-postgres
pandas
apache-airflow
apache-airflow[sentry]
apache-airflow-providers-celery
python-telegram-bot